+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.3.0 - 2016-11-06
-- New method findDocuments()
+- New method findDocuments() (requires CouchDB >= 2.0.0)
 - New property "duration" to report the execution time in milliseconds
 
 ## 1.2.1 - 2016-10-16